client front

+++ src/app/dashboard/client/client-list/client-list.component.ts
@@ -48,6 +48,7 @@ export class ClientListComponent implements OnInit {
     this.getClients();
   }
 
+  //Get List Clients
   getClients() {
     this.clientService.getClients().subscribe(
       (data) => {
@@ -63,30 +64,43 @@ export class ClientListComponent implements OnInit {
     )
   }
 
+  //delete ClientById
   deleteClient(id: string) {
-    this.clientService.deleteClient(id).subscribe(
-      (data) => {
-        this.clientList = this.clientList.filter((client: any) => {
-          return client.id != id
-        })
-        this.dataSource = new MatTableDataSource<any>(this.clientList);
-        this.dataSource.paginator = this.paginator;
-        this.dataSource.sort = this.sort;
-      },
-      (err) => {
-
+    this.swalService.confirmAlert(
+      'Delete Product',
+      'Are you sure you want to delete this product ?',
+      'Delete'
+    ).then((result: any) => {
+      if (result.value) {
+        this.clientService.deleteClient(id).subscribe(
+          (data) => {
+            this.clientList = this.clientList.filter((client: any) => {
+              return client.id != id
+            })
+            this.dataSource = new MatTableDataSource<any>(this.clientList);
+            this.dataSource.paginator = this.paginator;
+            this.dataSource.sort = this.sort;
+          },
+          (err) => {
+            console.log('error', err)
+            this.openSnackBar("server error", 'error')
+          }
+        )
       }
-    )
+    })
   }
 
+  openSnackBar(message: string, action: string) {
+    this._snackBar.open(message, action);
+  }
+
+  //Add Client
   openDialogCreateClient() {
     const dialogRef = this.dialog.open(UpdateClientComponent,
       {
         width: '80%',
         height: '60%',
-
       });
-
     dialogRef.disableClose = false;
     dialogRef.afterClosed().subscribe((res: any) => {
       if (!res)
@@ -99,6 +113,7 @@ export class ClientListComponent implements OnInit {
     });
   }
 
+  //Edit Client
   openDialogEditClient(element: any, index: number) {
     const dialogRef = this.dialog.open(UpdateClientComponent,
       {
@@ -117,4 +132,6 @@ export class ClientListComponent implements OnInit {
       this.dataSource.sort = this.sort;
     });
   }
+
+
 }
